Finishing up gremlin transferrence
Gremlins now exist between races. If you don't believe me, check the gremlinName for the player gremlin after selecting a gremlin for a race. I also fixed a bug where the racing gremlins didn't have an animator, so they couldn't move on the track. They're sideways now, but that can be fixed once the gremlin animations are added.